// frontend/src/shared/api/generated/model/getBoardgames200DataItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Boardgames tracker API
 * API для проекта Boardgames tracker
 * OpenAPI spec version: 1.0.0
 */
import type { GetBoardgames200DataItemTagsItem } from './getBoardgames200DataItemTagsItem';

export type GetBoardgames200DataItem = {
  readonly id: number;
  /** @nullable */
  bggId?: number | null;
  type: string;
  /** @maxLength 500 */
  originalName: string;
  /** @maxLength 500 */
  displayName: string;
  /** @nullable */
  complexity?: number | null;
  /** @nullable */
  minPlayers?: number | null;
  /** @nullable */
  maxPlayers?: number | null;
  /** @nullable */
  playingTime?: number | null;
  /** @nullable */
  minPlayTime?: number | null;
  /** @nullable */
  maxPlayTime?: number | null;
  /** @nullable */
  minAge?: number | null;
  /** @nullable */
  yearPublished?: number | null;
  /**
   * @maxLength 500
   * @nullable
   */
  imageUrl?: string | null;
  /**
   * @maxLength 500
   * @nullable
   */
  previewUrl?: string | null;
  isCustom: boolean;
  tags: GetBoardgames200DataItemTagsItem[];
};